# api/openapi.json
{
  "components": {
    "schemas": {
...
      "Error": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object",
        "x-go-type": "model.Error"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        },
        "required": [
          "error"
        ],
        "type": "object",
        "x-go-type": "model.ErrorEnvelope"
      },
...
      "List[github.com/jlbyh2o/llamaman/internal/api.DownloadDTO]": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DownloadDTO"
            },
            "type": "array"
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "total": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "type": "object",
        "x-go-type": "api.List[github.com/jlbyh2o/llamaman/internal/api.DownloadDTO]"
      },
...
    },
    "securitySchemes": {
      "csrfToken": {
        "description": "Double-submit echo of the non-HttpOnly `lm_csrf` cookie. Required on every non-GET of a session endpoint.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "sessionCookie": {
        "description": "`<session_id>.<secret>`; HttpOnly, SameSite=Lax, Path=/, Secure only over TLS. Only sha256(secret) is stored.",
        "in": "cookie",
        "name": "lm_session",
        "type": "apiKey"
      },
      "setupToken": {
        "description": "The one-time claim token, required from every non-loopback caller until the admin account exists (D38).",
        "in": "header",
        "name": "X-Setup-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "The Llama Man management API. Generated from the route registry in internal/api; do not hand-edit (DESIGN section 3, D43).",
    "title": "Llama Man",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/downloads": {
      "get": {
        "operationId": "listDownloads",
        "parameters": [
          {
            "description": "`active` is everything unfinished, `paused` included; `all` is the whole table, which is the receipt for what has landed on this disk.",
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "enum": [
                "active",
                "all"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/List[github.com/jlbyh2o/llamaman/internal/api.DownloadDTO]"
                }
              }
            },
            "description": "The downloads."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No valid admin session accompanied the request.",
            "x-error-codes": [
              "unauthorized"
            ]
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "This host has not been claimed yet. The SPA routes to the wizard on this code alone.",
            "x-error-codes": [
              "setup_required"
            ]
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An unhandled error. The message is a constant; the detail is in the journal.",
            "x-error-codes": [
              "internal_error"
            ]
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Downloads with their per-file progress.",
        "tags": [
          "downloads"
        ]
      },
      "post": {
        "operationId": "createDownload",
        "parameters": [
          {
            "description": "Optional replay key. A repeat within 10 minutes returns the original job with 200 instead of creating a second one; the same key with a different body inside the window is 422 idempotency_key_reused (D39/D65).",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDownloadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDownloadResponse"
                }
              }
            },
            "description": "An `Idempotency-Key` replay inside its window: the original download."
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDownloadResponse"
                }
              }
            },
            "description": "The download was queued. Watch `job_id`."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The Idempotency-Key header is not a short printable ASCII token.",
            "x-error-codes": [
              "idempotency_key_invalid"
            ]
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No valid admin session accompanied the request.",
            "x-error-codes": [
              "unauthorized"
            ]
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The repository is gated. `details` carries `repo` and `request_url`; access grants are browser-only on the Hub's side, so the UI links out.",
            "x-error-codes": [
              "hf_gated",
              "hf_private"
            ]
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "This model is already being downloaded (`details.download_id` names it), or the target filesystem cannot hold it (`details` carries the numbers).",
            "x-error-codes": [
              "download_exists",
              "insufficient_disk",
              "job_in_flight"
            ]
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The request names files this repository does not hold at this revision, a shard set the repository holds only part of, more than one quantization, or a projector choice that is ambiguous.",
            "x-error-codes": [
              "file_not_in_repo",
              "mmproj_ambiguous",
              "multiple_quants",
              "no_files_selected",
              "shard_set_incomplete"
            ]
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An unhandled error. The message is a constant; the detail is in the journal.",
            "x-error-codes": [
              "internal_error"
            ]
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The Hugging Face Hub could not be reached.",
            "x-error-codes": [
              "hf_unreachable"
            ]
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "This daemon has no primary cache root yet.",
            "x-error-codes": [
              "no_cache_root"
            ]
          }
        },
        "security": [
          {
            "csrfToken": [],
            "sessionCookie": []
          }
        ],
        "summary": "Queue a model download. The jobs, downloads, models and model_files rows are written in one transaction, so the job is the receipt.",
        "tags": [
          "downloads"
        ]
      }
    },
...
    "/api/v1/downloads/{id}/cancel": {
      "post": {
        "operationId": "cancelDownload",
        "parameters": [
          {
            "description": "Path segment.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Default true. False removes each `.incomplete` file as the transfer that owns it releases its handle.",
            "in": "query",
            "name": "keep_partial",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadDTO"
                }
              }
            },
            "description": "The canceled download."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No valid admin session accompanied the request.",
            "x-error-codes": [
              "unauthorized"
            ]
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The CSRF double-submit, Origin or Sec-Fetch-Site check failed.",
            "x-error-codes": [
              "csrf_failed"
            ]
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No download has this id.",
            "x-error-codes": [
              "not_found"
            ]
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The download has already finished.",
            "x-error-codes": [
              "download_not_pausable"
            ]
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "An unhandled error. The message is a constant; the detail is in the journal.",
            "x-error-codes": [
              "internal_error"
            ]
          }
        },
        "security": [
          {
            "csrfToken": [],
            "sessionCookie": []
          }
        ],
        "summary": "Cancel a download. The partial files are kept by default, so a retry resumes rather than starting over.",
        "tags": [
          "downloads"
        ]
      }
    },
...
  },
...
}